+++ lib/cancan.rb
@@ -1,23 +1,23 @@
 require "cancan/version"
 require 'cancan/ability'
 require 'cancan/rule'
-require 'cancan/helpers/utils/actions'
-require 'cancan/helpers/utils/authorization'
-require 'cancan/helpers/utils/id_param'
-require 'cancan/helpers/utils/name_methods'
-require 'cancan/helpers/utils/options'
-require 'cancan/helpers/utils/parent'
-require 'cancan/helpers/utils/resource_class'
-require 'cancan/helpers/utils/resource_class_parent'
-require 'cancan/helpers/base'
-require 'cancan/helpers/accessor'
-require 'cancan/helpers/authorizer'
-require 'cancan/helpers/builder'
-require 'cancan/helpers/finder'
-require 'cancan/helpers/loader'
-require 'cancan/helpers/inherited_resources_loader'
-require 'cancan/helpers/resource_class'
-require 'cancan/helpers/skipper'
+require 'cancan/concepts/utils/actions'
+require 'cancan/concepts/utils/authorization'
+require 'cancan/concepts/utils/id_param'
+require 'cancan/concepts/utils/name_methods'
+require 'cancan/concepts/utils/options'
+require 'cancan/concepts/utils/parent'
+require 'cancan/concepts/utils/resource_class'
+require 'cancan/concepts/utils/resource_class_parent'
+require 'cancan/concepts/base'
+require 'cancan/concepts/accessor'
+require 'cancan/concepts/authorizer'
+require 'cancan/concepts/builder'
+require 'cancan/concepts/finder'
+require 'cancan/concepts/loader'
+require 'cancan/concepts/inherited_resources_loader'
+require 'cancan/concepts/resource_class'
+require 'cancan/concepts/skipper'
 require 'cancan/controller_resource'
 require 'cancan/controller_additions'
 require 'cancan/model_additions'
