+++ lib/cancan/concepts/loader.rb
