+++ lib/cancan/concepts/utils/name_methods.rb
@@ -1,5 +1,5 @@
 module CanCan
-  module Helpers
+  module Concepts
     module Utils
       module NameMethods
