+++ lib/cancan/concepts/utils/resource_class_parent.rb
@@ -1,5 +1,5 @@
 module CanCan
-  module Helpers
+  module Concepts
     module Utils
       module ResourceClassParent
